Fix bugs

+++ app/schemas/route.py
@@ -7,18 +7,18 @@ class RouteOut(BaseModel):
     to: str = Field()
     from_: str = Field(serialization_alias="from")
     steps: List["StepOut"] = Field()
-    fullDistance: int = Field()
+    fullDistance: float = Field()
 
 
 class StepOut(BaseModel):
     plan: str = Field()
     way: List["WayOut"] = Field()
-    distance: int = Field()
+    distance: float = Field()
     pass
 
 
 class WayOut(BaseModel):
     id: str = Field()
-    x: int = Field()
-    y: int = Field()
+    x: float = Field()
+    y: float = Field()
     type: VertexType = Field()
